ReferenceSpace computes Jacobians: Seems OK -- TriEdgeBasis seems to be OK -- FunctionSpaceScalar and FunctionSpaceVector adapted with new ReferenceSpace::getJacobian

+++ FunctionSpace/TriEdgeBasis.h
@@ -20,7 +20,7 @@ class TriEdgeBasis: public BasisHierarchical1From{
   //! @param order The order of the Basis
   //!
   //! Returns a new Edge-Basis for Triangles of the given order
-  TriEdgeBasis(unsigned int order);
+  TriEdgeBasis(size_t order);
 
   //! Deletes this Basis
   //!
